pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "matplotlib>=3.10.3",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "requests>=2.32.4",
    "spotipy>=2.25.1",
    "trafilatura>=2.0.0",
    "yt-dlp>=2025.6.30",
]
